feat: use dhis config provider to setup app store
DHIS2-7257

Move APP_STORE_URL and APP_STORE_API_URL from runtime configurable
system settings to boot-time configured through the `dhis.conf` file.

This still uses the current app store (v1) as the default, and allows the
user to configure the new app store (v2) through `dhis.conf`.

```
appStoreUrl=https://default/url/to/appstore
appStoreApiUrl=https://default/url/to/appstore/api
```

This also cleans up the very old JSON based appstore as it was unused in
the codebase, and renames the "appstore2" package to "appstore".
