+++ dhis-2/dhis-services/dhis-service-administration/src/main/java/org/hisp/dhis/appstore/WebApp.java
@@ -28,28 +28,100 @@ package org.hisp.dhis.appstore;
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.hisp.dhis.appmanager.AppStatus;
+import org.hisp.dhis.appmanager.AppType;
+
+import java.util.Date;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
- * @author Lars Helge Overland
+ * Created by zubair@dhis2.org on 07.09.17.
  */
 public class WebApp
 {
+    private String id; // uid
+
     private String name;
-    
+
     private String description;
-    
-    private String imgLarge;
-    
-    private String developer;
-    
-    private List<WebAppVersion> versions = new ArrayList<>();
-    
+
+    private String owner;
+
+    private Date created;
+
+    private Date lastUpdated;
+
+    private String sourceUrl;
+
+    private AppStatus status;
+
+    private AppType appType;
+
+    private Developer developer;
+
+    private Set<AppVersion> versions = new HashSet<>();
+
+    private Set<Review> reviews = new HashSet<>();
+
+    private Set<ImageResource> images = new HashSet<>();
+
     public WebApp()
     {
+    }
+
+    @JsonProperty
+    public AppType getAppType()
+    {
+        return appType;
+    }
+
+    public void setAppType( AppType appType )
+    {
+        this.appType = appType;
+    }
+
+    @JsonProperty
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription( String description )
+    {
+        this.description = description;
+    }
+
+    @JsonProperty
+    public Developer getDeveloper() {
+        return developer;
+    }
+
+    public void setDeveloper( Developer developer )
+    {
+        this.developer = developer;
+    }
+
+    @JsonProperty
+    public String getId()
+    {
+        return id;
+    }
+
+    public void setId( String id )
+    {
+        this.id = id;
+    }
+
+    @JsonProperty
+    public Set<ImageResource> getImages()
+    {
+        return images;
+    }
+
+    public void setImages( Set<ImageResource> images )
+    {
+        this.images = images;
     }
 
     @JsonProperty
@@ -64,46 +136,79 @@ public class WebApp
     }
 
     @JsonProperty
-    public String getDescription()
+    public String getOwner()
     {
-        return description;
+        return owner;
     }
 
-    public void setDescription( String description )
+    public void setOwner( String owner )
     {
-        this.description = description;
-    }
-
-    @JsonProperty( value = "img_large" )
-    public String getImgLarge()
-    {
-        return imgLarge;
-    }
-
-    public void setImgLarge( String imgLarge )
-    {
-        this.imgLarge = imgLarge;
+        this.owner = owner;
     }
 
     @JsonProperty
-    public String getDeveloper()
+    public Set<Review> getReviews()
     {
-        return developer;
+        return reviews;
     }
 
-    public void setDeveloper( String developer )
+    public void setReviews( Set<Review> reviews )
     {
-        this.developer = developer;
+        this.reviews = reviews;
     }
 
     @JsonProperty
-    public List<WebAppVersion> getVersions()
+    public String getSourceUrl()
+    {
+        return sourceUrl;
+    }
+
+    public void setSourceUrl( String sourceUrl )
+    {
+        this.sourceUrl = sourceUrl;
+    }
+
+    @JsonProperty
+    public AppStatus getStatus()
+    {
+        return status;
+    }
+
+    public void setStatus( AppStatus status )
+    {
+        this.status = status;
+    }
+
+    @JsonProperty
+    public Set<AppVersion> getVersions()
     {
         return versions;
     }
 
-    public void setVersions( List<WebAppVersion> versions )
+    public void setVersions( Set<AppVersion> versions )
     {
         this.versions = versions;
+    }
+
+    @JsonProperty
+    public Date getCreated()
+    {
+        return created;
+    }
+
+    public void setCreated( Date created )
+    {
+        this.created = created;
+    }
+
+    @JsonProperty
+    public Date getLastUpdated()
+    {
+        return lastUpdated;
+    }
+
+    public void setLastUpdated( Date lastUpdated )
+    {
+        this.lastUpdated = lastUpdated;
     }
 }
