+++ dhis-2/dhis-services/dhis-service-setting/src/main/java/org/hisp/dhis/setting/SettingKey.java
@@ -143,10 +143,6 @@ public enum SettingKey
     CUSTOM_CSS( "keyCustomCss" ),
     CALENDAR( "keyCalendar", "iso8601", String.class ),
     DATE_FORMAT( "keyDateFormat", "yyyy-MM-dd", String.class ),
-    APP_STORE_URL( "appStoreUrl", "https://www.dhis2.org/appstore", String.class ),
-    APP_STORE_INDEX_URL( "keyAppStoreIndexUrl", "https://s3-eu-west-1.amazonaws.com/dhis2-appstore/appstore.json", String.class ),
-    APP_STORE( "keyAppStoreUrl", "https://play.dhis2.org/appstore/", String.class ),
-    APP_STORE_API_URL( "keyAppStoreApiUrl", "https://play.dhis2.org/appstore/api/apps", String.class ),
     STYLE( "keyStyle", "light_blue/light_blue.css", String.class ),
     REMOTE_INSTANCE_URL( "keyRemoteInstanceUrl", "", String.class ),
     REMOTE_INSTANCE_USERNAME( "keyRemoteInstanceUsername", "", String.class ),
